[refactor] Rename files and improve code

+++ src/V1/EmployeeContext/Employee/Application/DeleteEmployee/EmployeeDeletor.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Api\V1\EmployeeContext\Employee\Application\DeleteEmployee;
 
+use Api\V1\EmployeeContext\Employee\Domain\EmployeeNotExistsException;
 use Api\V1\EmployeeContext\Employee\Domain\EmployeeRepository;
 use Api\V1\EmployeeContext\Shared\Domain\Employee\EmployeeId;
 
@@ -11,13 +12,19 @@ final class EmployeeDeletor
 {
     public function __construct(
         private EmployeeRepository $repository
-    ) {
+    )
+    {
     }
 
     public function __invoke(
         EmployeeId $employeeId,
-    ): void {
+    ): void
+    {
         $employee = $this->repository->findById($employeeId);
+
+        if (null === $employee) {
+            throw new EmployeeNotExistsException();
+        }
 
         $employee->delete();
